Moved InterruptScopeGuard to separate file.
