+++ kernel/PageFrameAllocator.cpp
@@ -2,8 +2,6 @@
 #include <PageFrameAllocator.h>
 #include <Platform.h>
 
-PLATFORM_REQUIRED(MINOS_PLATFORM_X86_64)
-
 namespace Kernel
 {
     PageFrameAllocator globalPageManager;
