+++ kernel/drivers/CmosRTC.cpp
@@ -1,5 +1,6 @@
 #include <drivers/CmosRTC.h>
 #include <drivers/CPU.h>
+#include <InterruptScopeGuard.h>
 
 #define READ_RTC_REG(x, y) CPU::PortWrite8(PORT_CMOS_SELECT, x); y = CPU::PortRead8(PORT_CMOS_DATA);
